refactor: ScoreHistoryItem型の重複解消 (#852)
4つのコンポーネントに定義されたローカルSession interfaceを
共通のScoreHistoryItem型（types/index.ts）に統一。
